It is now possible to use ports defined in CAL files. See a summary of bug corrections/new features below. And enjoy!
Added an actor.owl ontology that imports xdf.owl (because the two formats have a lot in common). Corrected a bug in Cal grammar where identifiers could not have an underscore. Corrected a bug in calToXml.xslt that occurred when an actor would have only output ports. Solved a bug in XsltTransformer by outputting the result of the transformation directly to a document, not to its root element.

git-svn-id: http://svn.code.sf.net/p/graphiti-editor/code@192 04dbf5ff-8551-0410-9f1f-ce2c5325ca72